requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
chromadb==1.3.5
langchain==1.0.8
langchain-openai==1.0.3
langchain-community==0.4.1
pydantic==2.11.1
python-dotenv==1.2.1
python-multipart==0.0.20
httpx==0.28.3
openai==1.71.0
